requirements.txt
nba_api
pandas
matplotlib
python-dotenv
openai
requests
mplcursors 